requirements.txt
# =============================================================================
# HandsDB Requirements
# Generated: 2026-02-03
# Python >= 3.10
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.1

# -----------------------------------------------------------------------------
# Template Engine
# -----------------------------------------------------------------------------
jinja2==3.1.3
MarkupSafe==3.0.3

# -----------------------------------------------------------------------------
# Database Drivers
# -----------------------------------------------------------------------------
pyodbc==5.0.1
psycopg2-binary==2.9.11
PyMySQL==1.1.2
oracledb==3.4.2
SQLAlchemy==2.0.25
greenlet==3.3.1

# -----------------------------------------------------------------------------
# Data Validation
# -----------------------------------------------------------------------------
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
annotated-types==0.7.0
typing_extensions==4.15.0

# -----------------------------------------------------------------------------
# Security & Authentication
# -----------------------------------------------------------------------------
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==46.0.4
cffi==2.0.0
pycparser==3.0
ecdsa==0.19.1
pyasn1==0.6.2
rsa==4.9.1

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
python-multipart==0.0.6
click==8.3.1
colorama==0.4.6
six==1.17.0
PyYAML==6.0.3

# -----------------------------------------------------------------------------
# Async Support
# -----------------------------------------------------------------------------
anyio==4.12.1
sniffio==1.3.1
h11==0.16.0
httptools==0.7.1
watchfiles==1.1.1
websockets==16.0

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx==0.26.0
httpcore==1.0.9
certifi==2026.1.4
idna==3.11

# -----------------------------------------------------------------------------
# Excel
# -----------------------------------------------------------------------------
openpyxl==3.1.5
et_xmlfile==2.0.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pluggy==1.6.0
iniconfig==2.3.0
packaging==26.0
pymssql
